+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- support legacy monorepo format workspaces: { packages: [] }
+
 ## [1.0.0] - 2026-03-16
 
 Initial release